+++ 2D/src/Engine.h
@@ -28,8 +28,8 @@ private:
 	static constexpr int DEFAULT_WIDTH{ 1280 };
 	static constexpr int DEFAULT_HEIGHT{ 720 };
 	static constexpr int FPS { 60 };
-	size_t mWindowWidth{ DEFAULT_WIDTH };
-	size_t mWindowHeight{ DEFAULT_HEIGHT };
+	int mWindowWidth{ DEFAULT_WIDTH };
+	int mWindowHeight{ DEFAULT_HEIGHT };
 	void Update();
 	void Input();
 	sf::Clock clock;
